introduce rules and variables in separate dedicated files
