+++ src/components/MessageList/Message/Message.tsx
@@ -2,35 +2,10 @@ import * as React from 'react'
 import PropTypes from 'prop-types'
 
 import cx from 'classnames'
-import { createComponent, customPropTypes, getElementType } from '../../lib'
+import { createComponent, customPropTypes, getElementType } from '../../../lib'
 
-const messageRules = ({ isMine, variables }) => ({
-  container: {
-    display: 'flex',
-    margin: '1rem', // TODO is temporary hardcoded value
-    justifyContent: (isMine && 'flex-end') || 'flex-start',
-  },
-  text: {
-    display: 'inline-block',
-    backgroundColor: (isMine && variables.messageColor.mine) || variables.messageColor.incoming,
-    maxWidth: variables.messageWidth,
-
-    // TODO - the following ones are temporary
-    margin: 0,
-    padding: '1rem',
-    borderRadius: '0.3rem',
-    color: 'rgb(64,64,64)',
-    fontFamily: 'Segoe UI, sans-serif',
-  },
-})
-
-const messageVariables = {
-  messageWidth: '80%',
-  messageColor: {
-    mine: '#E0E0ED',
-    incoming: 'rgba(0,0,0,0.1)',
-  },
-}
+import messageRules from './messageRules'
+import messageVariables from './messageVariables'
 
 const Message: any = props => {
   const { content, styles, className } = props
