+++ src/components/MessageList/Message/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Message'
