+++ src/components/MessageList/Message/messageVariables.ts
@@ -1,0 +1,9 @@
+const messageVariables = {
+  messageWidth: '80%',
+  messageColor: {
+    mine: '#E0E0ED',
+    incoming: 'rgba(0,0,0,0.1)',
+  },
+}
+
+export default messageVariables
