+++ src/components/MessageList/messageListRules.ts
@@ -1,0 +1,10 @@
+const messageListRules = () => ({
+  root: {
+    listStyle: 'none',
+    display: 'block',
+    padding: 0,
+    margin: 0,
+  },
+})
+
+export default messageListRules
